Refactor: separate entry-related types into different files.
